+++ actions/src/robocorp/actions/_protocols.py
@@ -3,6 +3,7 @@ from typing import Any, Callable, Sequence
 
 from robocorp.tasks import ITask as _ITask
 from robocorp.tasks import Status as _Status
+from robocorp.tasks._protocols import TasksListTaskTypedDict
 
 
 class IAction(_ITask, typing.Protocol):
@@ -13,3 +14,7 @@ Status = _Status
 
 IActionCallback = Callable[[IAction], Any]
 IActionsCallback = Callable[[Sequence[IAction]], Any]
+
+
+class ActionsListActionTypedDict(TasksListTaskTypedDict):
+    pass