+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- `action-server package metadata` now includes information on the secrets required
+  for each action.
+
 ## 0.2.1 - 2024-04-04
 
 - `action-server package build` no longer includes the `.zip` being created in the